fix broken debug build - missing symbols when using constexpr

+++ libs/ledger/src/chain/main_chain.cpp
@@ -229,7 +229,7 @@ bool MainChain::RemoveBlock(BlockHash hash)
  */
 MainChain::Blocks MainChain::GetHeaviestChain(uint64_t limit) const
 {
-  limit = std::min(limit, MainChain::UPPER_BOUND);
+  limit = std::min(limit, uint64_t(MainChain::UPPER_BOUND));
   MilliTimer myTimer("MainChain::HeaviestChain");
 
   FETCH_LOCK(lock_);
@@ -299,7 +299,7 @@ MainChain::Blocks MainChain::GetChainPreceding(BlockHash start, uint64_t limit) 
 bool MainChain::GetPathToCommonAncestor(Blocks &blocks, BlockHash tip, BlockHash node,
                                         uint64_t limit, BehaviourWhenLimit behaviour) const
 {
-  limit = std::min(limit, MainChain::UPPER_BOUND);
+  limit = std::min(limit, uint64_t(MainChain::UPPER_BOUND));
   MilliTimer myTimer("MainChain::GetPathToCommonAncestor", 500);
 
   FETCH_LOCK(lock_);
@@ -457,7 +457,7 @@ MainChain::BlockHashSet MainChain::GetMissingTips() const
  */
 MainChain::BlockHashs MainChain::GetMissingBlockHashes(uint64_t limit) const
 {
-  limit = std::min(limit, MainChain::UPPER_BOUND);
+  limit = std::min(limit, uint64_t(MainChain::UPPER_BOUND));
   FETCH_LOCK(lock_);
 
   BlockHashs results;
